Move NotificationPage
