+++ airbyte-webapp/src/pages/SettingsPage/pages/NotificationPage/index.tsx
@@ -1,0 +1,3 @@
+import NotificationPage from "./NotificationPage";
+
+export default NotificationPage;
